pyproject.toml
[tool.poetry]
name = "aihack"
version = "0.1.0"
description = "Privacy-first AI development assistant"
authors = ["Samuel K. Norton <samuelknorton@gmail.com>"]
readme = "README.md"

packages = [
  { include = "aihack", from = "src"  }
]


[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.0"
rich = "^13.0.0"
httpx = "^0.25.0"
# AI Libraries - use >= for faster updates in this fast-moving space
anthropic = ">=0.62.0,<1.0.0"
google-generativeai = ">=0.3.0,<1.0.0"
# Stable libraries
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
gitpython = "^3.1.0"
pyyaml = "^6.0.1"
tree-sitter = "^0.20.0"
watchdog = "^3.0.0"
tomli = "^2.2.1"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
ah = "aihack.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["aihack"]

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["anthropic.*", "google.generativeai.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "examples.*",
    "aihack.tests.scripts.*",
    "aihack.core.click",
    "aihack.config"
]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["aihack.cli.*", "aihack.tests.*"]
disallow_untyped_decorators = false

# ruff configuration
[tool.ruff]
exclude = [
    "examples/",
    "src/aihack/tests/scripts/",
]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["src/aihack/tests", "tests"]
python_files = "test*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
